Fix another bug preventing save when you click on a new due date more than once.
Also, remove the Fable warning for 'String.Contains: second argument is ignored' by not passing in the StringComparer. Apparently it's just not supported for Contains. Note to self: pay attention to warnings in future because they might prevent bugs!

(The bug this warning was about has already been fixed by just lowercasing everything.)